Move RigFlowDiagTimeStepResult to separate file
