+++ ApplicationLibCode/ReservoirDataModel/CMakeLists_files.cmake
@@ -39,6 +39,7 @@ set(SOURCE_GROUP_HEADER_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RigFlowDiagResults.h
     ${CMAKE_CURRENT_LIST_DIR}/RigFlowDiagSolverInterface.h
     ${CMAKE_CURRENT_LIST_DIR}/RigFlowDiagStatCalc.h
+    ${CMAKE_CURRENT_LIST_DIR}/RigFlowDiagTimeStepResult.h
     ${CMAKE_CURRENT_LIST_DIR}/RigFlowDiagVisibleCellsStatCalc.h
     ${CMAKE_CURRENT_LIST_DIR}/RigFlowDiagDefines.h
     ${CMAKE_CURRENT_LIST_DIR}/RigFormationNames.h
@@ -128,6 +129,7 @@ set(SOURCE_GROUP_SOURCE_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RigFlowDiagResults.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RigFlowDiagSolverInterface.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RigFlowDiagStatCalc.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/RigFlowDiagTimeStepResult.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RigFlowDiagVisibleCellsStatCalc.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RigFlowDiagDefines.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RigFormationNames.cpp
