+++ ApplicationLibCode/ReservoirDataModel/RigFlowDiagSolverInterface.cpp
@@ -45,75 +45,6 @@
 
 #include "cvfTrace.h"
 
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-RigFlowDiagTimeStepResult::RigFlowDiagTimeStepResult( size_t activeCellCount )
-    : m_activeCellCount( activeCellCount )
-{
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void RigFlowDiagTimeStepResult::setTracerTOF( const std::string&                       tracerName,
-                                              RigFlowDiagResultAddress::PhaseSelection phaseSelection,
-                                              const std::map<int, double>&             cellValues )
-{
-    std::set<std::string> tracers;
-    tracers.insert( tracerName );
-
-    RigFlowDiagResultAddress resAddr( RigFlowDiagDefines::tofResultName().toStdString(), phaseSelection, tracers );
-
-    addResult( resAddr, cellValues );
-
-    std::vector<double>& activeCellValues = m_nativeResults[resAddr];
-    for ( double& val : activeCellValues )
-    {
-        val = val * 1.15741e-5; // days pr second. Converting to days
-    }
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void RigFlowDiagTimeStepResult::setTracerFraction( const std::string&                       tracerName,
-                                                   RigFlowDiagResultAddress::PhaseSelection phaseSelection,
-                                                   const std::map<int, double>&             cellValues )
-{
-    std::set<std::string> tracers;
-    tracers.insert( tracerName );
-
-    addResult( RigFlowDiagResultAddress( RigFlowDiagDefines::cellFractionResultName().toStdString(), phaseSelection, tracers ), cellValues );
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void RigFlowDiagTimeStepResult::setInjProdWellPairFlux( const std::string&               injectorTracerName,
-                                                        const std::string&               producerTracerName,
-                                                        const std::pair<double, double>& injProdFluxes )
-{
-    m_injProdWellPairFluxes[std::make_pair( injectorTracerName, producerTracerName )] = injProdFluxes;
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void RigFlowDiagTimeStepResult::addResult( const RigFlowDiagResultAddress& resAddr, const std::map<int, double>& cellValues )
-{
-    std::vector<double>& activeCellValues = m_nativeResults[resAddr];
-
-    CVF_ASSERT( activeCellValues.empty() );
-
-    activeCellValues.resize( m_activeCellCount, HUGE_VAL );
-
-    for ( const auto& pairIt : cellValues )
-    {
-        activeCellValues[pairIt.first] = pairIt.second;
-    }
-}
-
 class RigOpmFlowDiagStaticData
 {
 public:
